Branched trunk to stable version 2.0, so stable now includes all the Clarin changes to date.

+++ src/nl/mpi/arbil/LinorgFavourites.java
@@ -1,10 +1,8 @@
 package nl.mpi.arbil;
 
 import nl.mpi.arbil.data.ImdiTreeObject;
-import nl.mpi.arbil.data.ImdiSchema;
+import nl.mpi.arbil.MetadataFile.MetadataReader;
 import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URI;
 import java.util.Enumeration;
 import java.util.Hashtable;
@@ -29,18 +27,24 @@ public class LinorgFavourites {
     }
 
     // this will load any favourites in the old format and delete the old format file
-    public void loadOldFormatFavourites() {
+    public void convertOldFormatLocationLists() {
         try {
-            Vector<String> locationsList = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("locationsList");
+            File oldLocationsFile = new File(LinorgSessionStorage.getSingleInstance().storageDirectory, "locationsList");
             File oldFavouritesFile = new File(LinorgSessionStorage.getSingleInstance().storageDirectory, "selectedFavourites");
-            if (oldFavouritesFile.exists()) {
-                Vector<String> userFavouritesStrings = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("selectedFavourites");
-                locationsList.addAll(userFavouritesStrings);
-                LinorgSessionStorage.getSingleInstance().saveObject(locationsList, "locationsList");
-                oldFavouritesFile.deleteOnExit();
+            if (oldLocationsFile.exists()) {
+                Vector<String> locationsList = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("locationsList");
+                if (oldFavouritesFile.exists()) {
+                    Vector<String> userFavouritesStrings = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("selectedFavourites");
+                    locationsList.addAll(userFavouritesStrings);
+                    LinorgSessionStorage.getSingleInstance().saveStringArray("locationsList", locationsList.toArray(new String[]{}));
+                    oldFavouritesFile.deleteOnExit();
+                }
+                if (null == LinorgSessionStorage.getSingleInstance().loadStringArray("locationsList")) {
+                    LinorgSessionStorage.getSingleInstance().saveStringArray("locationsList", locationsList.toArray(new String[]{}));
+                }
             }
         } catch (Exception ex) {
-            System.out.println("load old format favourites failed: " + ex.getMessage());
+            GuiHelper.linorgBugCatcher.logError(ex);
         }
     }
 
@@ -76,7 +80,12 @@ public class LinorgFavourites {
 
     private void addAsFavourite(URI imdiUri) {
         try {
-            URI copiedFileURI = copyToFavouritesDirectory(imdiUri);
+            URI baseUri = new URI(imdiUri.toString().split("#")[0]);
+            String fileSuffix = imdiUri.getPath().substring(imdiUri.getPath().lastIndexOf("."));
+            File destinationFile = File.createTempFile("fav-", fileSuffix, LinorgSessionStorage.getSingleInstance().getFavouritesDir());
+            ImdiTreeObject.getMetadataUtils(baseUri.toString()).copyMetadataFile(baseUri, destinationFile, null, true);
+
+            URI copiedFileURI = destinationFile.toURI();
             // creating a uri with separate parameters could cause the url to be reencoded
             // hence this has been converted to use the string URI constuctor
 //            URI favouriteUri = new URI(copiedFileURI.getScheme(), copiedFileURI.getUserInfo(), copiedFileURI.getHost(), copiedFileURI.getPort(), copiedFileURI.getPath(), copiedFileURI.getQuery(),
@@ -97,29 +106,6 @@ public class LinorgFavourites {
             TreeHelper.getSingleInstance().applyRootLocations();
         } catch (Exception ex) {
             GuiHelper.linorgBugCatcher.logError(ex);
-        }
-    }
-
-    private URI copyToFavouritesDirectory(URI imdiUri) throws MalformedURLException, IOException {
-        mpi.util.OurURL inUrlLocal = new mpi.util.OurURL(imdiUri.toURL());
-        File destinationFile = File.createTempFile("fav-", ".imdi", LinorgSessionStorage.getSingleInstance().getFavouritesDir());
-        mpi.util.OurURL destinationUrl = new mpi.util.OurURL(destinationFile.toURL());
-
-        org.w3c.dom.Document nodDom = ImdiTreeObject.api.loadIMDIDocument(inUrlLocal, false);
-        if (nodDom == null) {
-            GuiHelper.linorgBugCatcher.logError(new Exception(ImdiTreeObject.api.getMessage()));
-            LinorgWindowManager.getSingleInstance().addMessageDialogToQueue("Error reading via the IMDI API", "Save Favourite");
-            return null;
-        } else {
-            mpi.imdi.api.IMDILink[] links = ImdiTreeObject.api.getIMDILinks(nodDom, inUrlLocal, mpi.imdi.api.WSNodeType.UNKNOWN);
-            if (links != null) {
-                for (mpi.imdi.api.IMDILink currentLink : links) {
-                    ImdiTreeObject.api.changeIMDILink(nodDom, destinationUrl, currentLink);
-                }
-            }
-            boolean removeIdAttributes = false;
-            ImdiTreeObject.api.writeDOM(nodDom, destinationFile, removeIdAttributes);
-            return destinationFile.toURI();
         }
     }
 
@@ -155,9 +141,9 @@ public class LinorgFavourites {
                 if (targetIsCorpus && !currentFavouritesObject.isImdiChild()) {
                     addThisFavourites = true;
                 } else if (targetIsSession && currentFavouritesObject.isImdiChild()) {
-                    addThisFavourites = ImdiSchema.getSingleInstance().nodeCanExistInNode(targetImdiObject, currentFavouritesObject);
+                    addThisFavourites = MetadataReader.getSingleInstance().nodeCanExistInNode(targetImdiObject, currentFavouritesObject);
                 } else if (targetIsImdiChild && currentFavouritesObject.isImdiChild()) {
-                    addThisFavourites = ImdiSchema.getSingleInstance().nodeCanExistInNode(targetImdiObject, currentFavouritesObject);
+                    addThisFavourites = MetadataReader.getSingleInstance().nodeCanExistInNode(targetImdiObject, currentFavouritesObject);
                 }
                 if (addThisFavourites) {
 //                    System.out.println("adding: " + currentFavouritesObject);
@@ -188,9 +174,9 @@ public class LinorgFavourites {
         System.out.println("getNodeType: \nfavouriteXmlPath: " + favouriteXmlPath + "\ntargetXmlPath:" + targetXmlPath);
         String returnValue;
         if (favouriteImdiObject.isSession()) {
-            returnValue = ImdiSchema.imdiPathSeparator + "METATRANSCRIPT" + ImdiSchema.imdiPathSeparator + "Session";
+            returnValue = MetadataReader.imdiPathSeparator + "METATRANSCRIPT" + MetadataReader.imdiPathSeparator + "Session";
         } else if (favouriteImdiObject.isCorpus()) {
-            returnValue = ImdiSchema.imdiPathSeparator + "METATRANSCRIPT" + ImdiSchema.imdiPathSeparator + "Corpus";
+            returnValue = MetadataReader.imdiPathSeparator + "METATRANSCRIPT" + MetadataReader.imdiPathSeparator + "Corpus";
         } else if (favouriteImdiObject.isImdiChild()) {
             if (targetXmlPath == null) {
                 returnValue = favouriteXmlPath.replaceAll("\\(\\d*?\\)$", "");
@@ -220,38 +206,38 @@ public class LinorgFavourites {
         return returnValue;
     }
 
-    public void mergeFromFavourite(ImdiTreeObject targetImdiObject, ImdiTreeObject favouriteImdiObject, boolean overwriteValues) {
-//        System.out.println("mergeFromFavourite: " + addedNodeUrl + " : " + imdiTemplateUrl);
-        Hashtable<String, ImdiField[]> targetFieldsHash = targetImdiObject.getFields();
-        for (Enumeration<ImdiField[]> favouriteFieldEnum = favouriteImdiObject.getFields().elements(); favouriteFieldEnum.hasMoreElements();) {
-            ImdiField[] currentFavouriteFields = favouriteFieldEnum.nextElement();
-            if (currentFavouriteFields.length > 0) {
-                ImdiField[] targetNodeFields = targetFieldsHash.get(currentFavouriteFields[0].getTranslateFieldName());
-
-                System.out.println("TranslateFieldName: " + currentFavouriteFields[0].getTranslateFieldName());
-                System.out.println("targetImdiObjectLoading: " + targetImdiObject.isLoading());
-                if (targetNodeFields != null) {
-                    System.out.println("copy fields");
-                    for (int fieldCounter = 0; fieldCounter < currentFavouriteFields.length; fieldCounter++) {
-                        ImdiField currentField;
-                        if (targetNodeFields.length > fieldCounter) {
-                            // copy to the exisiting fields
-                            currentField = targetNodeFields[fieldCounter];
-                            currentField.setFieldValue(currentFavouriteFields[fieldCounter].getFieldValue(), false, false);
-                        } else {
-                            // add sub nodes if they dont already exist
-                            currentField = new ImdiField(targetImdiObject, currentFavouriteFields[fieldCounter].xmlPath, "");
-                            currentField.setFieldValue(currentFavouriteFields[fieldCounter].getFieldValue(), false, true); // this is done separatly to trigger the needs save to disk flag
-                            targetImdiObject.addField(currentField);
-//                            currentField.fieldNeedsSaveToDisk = true;
-                        }
-                        String currentLanguageId = currentFavouriteFields[fieldCounter].getLanguageId();
-                        if (currentLanguageId != null) {
-                            currentField.setLanguageId(currentLanguageId, false, true);
-                        }
-                    }
-                }
-            }
-        }
-    }
+//    public void mergeFromFavourite(ImdiTreeObject targetImdiObject, ImdiTreeObject favouriteImdiObject, boolean overwriteValues) {
+////        System.out.println("mergeFromFavourite: " + addedNodeUrl + " : " + imdiTemplateUrl);
+//        Hashtable<String, ImdiField[]> targetFieldsHash = targetImdiObject.getFields();
+//        for (Enumeration<ImdiField[]> favouriteFieldEnum = favouriteImdiObject.getFields().elements(); favouriteFieldEnum.hasMoreElements();) {
+//            ImdiField[] currentFavouriteFields = favouriteFieldEnum.nextElement();
+//            if (currentFavouriteFields.length > 0) {
+//                ImdiField[] targetNodeFields = targetFieldsHash.get(currentFavouriteFields[0].getTranslateFieldName());
+//
+//                System.out.println("TranslateFieldName: " + currentFavouriteFields[0].getTranslateFieldName());
+//                System.out.println("targetImdiObjectLoading: " + targetImdiObject.isLoading());
+//                if (targetNodeFields != null) {
+//                    System.out.println("copy fields");
+//                    for (int fieldCounter = 0; fieldCounter < currentFavouriteFields.length; fieldCounter++) {
+//                        ImdiField currentField;
+//                        if (targetNodeFields.length > fieldCounter) {
+//                            // copy to the exisiting fields
+//                            currentField = targetNodeFields[fieldCounter];
+//                            currentField.setFieldValue(currentFavouriteFields[fieldCounter].getFieldValue(), false, false);
+//                        } else {
+//                            // add sub nodes if they dont already exist
+//                            currentField = new ImdiField(0, targetImdiObject, currentFavouriteFields[fieldCounter].xmlPath, "", 0); // this is not correct but this section should be simplified asap
+//                            currentField.setFieldValue(currentFavouriteFields[fieldCounter].getFieldValue(), false, true); // this is done separatly to trigger the needs save to disk flag
+//                            targetImdiObject.addField(currentField);
+////                            currentField.fieldNeedsSaveToDisk = true;
+//                        }
+//                        String currentLanguageId = currentFavouriteFields[fieldCounter].getLanguageId();
+//                        if (currentLanguageId != null) {
+//                            currentField.setLanguageId(currentLanguageId, false, true);
+//                        }
+//                    }
+//                }
+//            }
+//        }
+//    }
 }
